# Do not emit full-drain event if Buffer is closed during drain event