--- a/tests/BufferTest.php
+++ b/tests/BufferTest.php
@@ -130,8 +130,6 @@ class BufferTest extends TestCase
      */
     public function testWriteInDrain()
     {
-        $writeStreams = array();
-
         $stream = fopen('php://temp', 'r+');
         $loop = $this->createWriteableLoopMock();
         $loop->preventWrites = true;
@@ -152,6 +150,47 @@ class BufferTest extends TestCase
 
         fseek($stream, 0);
         $this->assertSame("foo\nbar\n", stream_get_contents($stream));
+    }
+
+    /**
+     * @covers React\Stream\Buffer::write
+     * @covers React\Stream\Buffer::handleWrite
+     */
+    public function testDrainAndFullDrainAfterWrite()
+    {
+        $stream = fopen('php://temp', 'r+');
+        $loop = $this->createLoopMock();
+
+        $buffer = new Buffer($stream, $loop);
+        $buffer->softLimit = 2;
+
+        $buffer->on('drain', $this->expectCallableOnce());
+        $buffer->on('full-drain', $this->expectCallableOnce());
+
+        $buffer->write("foo");
+        $buffer->handleWrite();
+    }
+
+    /**
+     * @covers React\Stream\Buffer::write
+     * @covers React\Stream\Buffer::handleWrite
+     */
+    public function testCloseDuringDrainWillNotEmitFullDrain()
+    {
+        $stream = fopen('php://temp', 'r+');
+        $loop = $this->createLoopMock();
+
+        $buffer = new Buffer($stream, $loop);
+        $buffer->softLimit = 2;
+
+        // close buffer on drain event => expect close event, but no full-drain after
+        $buffer->on('drain', $this->expectCallableOnce());
+        $buffer->on('drain', array($buffer, 'close'));
+        $buffer->on('close', $this->expectCallableOnce());
+        $buffer->on('full-drain', $this->expectCallableNever());
+
+        $buffer->write("foo");
+        $buffer->handleWrite();
     }
 
     /**