The `Buffer` can be closed during the `drain` event, do not emit any further events in this case.
This is a very subtle bug and unlikely to ever have caused any real issues, but let's fix this for consistency.

Also, while we're at this, let's avoid some unnecessary function calls and improve performance a bit :)
Do not expect any major performance improvements here.

Running the examples/benchmark-throughput.php script, this change alone increases performance from ~100 MiB/s to ~110 MiB/s on my (rather mediocre) test system, using PHP 7.0.8.
Combined with #53 and #54, it increases performance from ~100 MiB/s to ~1950 MiB/s.

If you want to review, consider looking at the individual commits, this should make the changes pretty obvious.